refactor: Station 을 삭제하는 기능 수정

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -6,7 +6,9 @@ import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.request.StationRequestDto;
+import wooteco.subway.exception.CanNotDeleteException;
 import wooteco.subway.exception.DuplicateStationNameException;
+import wooteco.subway.repository.dao.JdbcSectionDao;
 import wooteco.subway.repository.dao.StationDao;
 import wooteco.subway.repository.entity.StationEntity;
 
@@ -14,9 +16,12 @@ import wooteco.subway.repository.entity.StationEntity;
 public class StationService {
 
     private final StationDao stationDao;
+    // 인터페이스로 바꿔야함
+    private final JdbcSectionDao sectionDao;
 
-    public StationService(final StationDao stationDao) {
+    public StationService(final StationDao stationDao, final JdbcSectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public Station register(final StationRequestDto stationRequestDto) {
@@ -42,6 +47,9 @@ public class StationService {
     }
 
     public void remove(final Long id) {
+        if (sectionDao.findByStationId(id).size() != 0) {
+            throw new CanNotDeleteException();
+        }
         stationDao.deleteById(id);
     }
 }
